+++ lib/src/local_working_copy.rs
@@ -798,6 +798,7 @@ impl TreeState {
             base_ignores,
             fsmonitor_settings,
             progress,
+            start_tracking_matcher,
             max_new_file_size,
         } = options;
 
@@ -835,6 +836,7 @@ impl TreeState {
             };
             self.visit_directory(
                 &matcher,
+                start_tracking_matcher,
                 &current_tree,
                 tree_entries_tx,
                 file_states_tx,
@@ -908,6 +910,7 @@ impl TreeState {
     fn visit_directory(
         &self,
         matcher: &dyn Matcher,
+        start_tracking_matcher: &dyn Matcher,
         current_tree: &MergedTree,
         tree_entries_tx: Sender<(RepoPathBuf, MergedTreeValue)>,
         file_states_tx: Sender<(RepoPathBuf, FileState)>,
@@ -1008,6 +1011,14 @@ impl TreeState {
                                 }
                             }
                         }
+                    } else if start_tracking_matcher.visit(&path).is_nothing()
+                        && file_states.is_empty()
+                    {
+                        // Don't visit subdirectory if it's not already tracked
+                        // and we should not start tracking it. The user might
+                        // have a huge target/ directory that they have not yet
+                        // added to their ignore patterns.
+                        // TODO: Report this directory to the caller
                     } else {
                         let directory_to_visit = DirectoryToVisit {
                             dir: path,
@@ -1017,6 +1028,7 @@ impl TreeState {
                         };
                         self.visit_directory(
                             matcher,
+                            start_tracking_matcher,
                             current_tree,
                             tree_entries_tx.clone(),
                             file_states_tx.clone(),
@@ -1034,8 +1046,12 @@ impl TreeState {
                         && git_ignore.matches(path.as_internal_file_string())
                     {
                         // If it wasn't already tracked and it matches
-                        // the ignored paths, then
-                        // ignore it.
+                        // the ignored paths, then ignore it.
+                    } else if maybe_current_file_state.is_none()
+                        && !start_tracking_matcher.matches(&path)
+                    {
+                        // Leave the file untracked
+                        // TODO: Report this path to the caller
                     } else {
                         let metadata = entry.metadata().map_err(|err| SnapshotError::Other {
                             message: format!("Failed to stat file {}", entry.path().display()),
@@ -1043,6 +1059,7 @@ impl TreeState {
                         })?;
                         if maybe_current_file_state.is_none() && metadata.len() > max_new_file_size
                         {
+                            // TODO: Maybe leave the file untracked instead
                             return Err(SnapshotError::NewFileTooLarge {
                                 path: entry.path().clone(),
                                 size: HumanByteSize(metadata.len()),